Fixed devNonce issue

+++ src/radio/radio.h
@@ -156,16 +156,16 @@ struct Radio_s
      */
 	bool ( *IsChannelFree )( uint32_t freq, uint32_t rxBandwidth, int16_t rssiThresh, uint32_t maxCarrierSenseTime );
 	/*!
-     * \brief Generates a 32 bits random value based on the RSSI readings
+     * \brief Generates a 16 bits random value based on the RSSI readings
      *
      * \remark This function sets the radio in LoRa modem mode and disables
      *         all interrupts.
      *         After calling this function either Radio.SetRxConfig or
      *         Radio.SetTxConfig functions must be called.
      *
-     * \retval randomValue    32 bits random value
+     * \retval randomValue    16 bits random value
      */
-	uint32_t ( *Random )( void );
+	uint16_t ( *Random )( void );
 	/*!
      * \brief Sets the reception parameters
      *
